Fixed interpreters & tests.

+++ src/sip/worker/profile_timer_store.h
@@ -30,7 +30,7 @@ public:
 	 * @param opcode_operands opcode and its operand blocks' shapes & sizes
 	 * @param time_count_pair pair of total time & number of times opcode was called
 	 */
-	void save_to_store(const ProfileTimer::Key& opcode_operands, const std::pair<long, long>& time_count_pair);
+	void save_to_store(const ProfileTimer::Key& opcode_operands, const std::pair<double, long>& time_count_pair);
 
 	/**
 	 * Retrieves from a persistent database, the profiled total time & count
@@ -38,10 +38,10 @@ public:
 	 * @param opcode_operands
 	 * @return
 	 */
-	std::pair<long, long> get_from_store(const ProfileTimer::Key& opcode_operands) const;
+	std::pair<double, long> get_from_store(const ProfileTimer::Key& opcode_operands) const;
 
 
-	typedef std::map<sip::ProfileTimer::Key, std::pair<long, long> > ProfileStoreMap_t;
+	typedef std::map<sip::ProfileTimer::Key, std::pair<double, long> > ProfileStoreMap_t;
 	typedef ProfileStoreMap_t::iterator ProfileStoreMapIterator_t;
